+++ tests/review.rs
@@ -41,6 +41,27 @@ fn test_label_boolean() -> Result<(), rexpect::errors::Error> {
 
 #[cfg(unix)]
 #[test]
+fn test_label_uri() -> Result<(), rexpect::errors::Error> {
+    let dir = "test-resources/label-uri";
+    common::remove_sink(dir).unwrap();
+    let mut p = common::spawn(dir, vec!["review", "label"], 1661192610, 400)?;
+    p.exp_string("acute toxicity? [Yes/No/Skip]")?;
+    p.send_line("y")?;
+    p.exp_string("eye irritation? [Yes/No/Skip]")?;
+    p.send_line("n")?;
+    p.exp_string("substance")?;
+    p.exp_string("1. \"sodium laureth sulfate\"")?;
+    p.exp_string("7. Skip Question")?;
+    p.exp_string("?")?;
+    p.send_line("1")?;
+    p.exp_string("acute toxicity? [Yes/No/Skip]")?;
+    p.send_control('c')?;
+    common::check_sink(dir).unwrap();
+    Ok(())
+}
+
+#[cfg(unix)]
+#[test]
 fn test_label_json_schema() -> Result<(), rexpect::errors::Error> {
     let dir = "test-resources/label-json-schema";
     common::remove_sink(dir).unwrap();